OvmfPkg: Add back QemuFlashFvbServicesRuntimeDxe
This allows flash variable store to be used by bhyve, along with
necessary changes in bhyve(8) and vmm(4).

Note that this change is not without peril. Specifically, the
probing behaviour in QemuFlashDetected() will cause a paging fault
and vm crash when using the combined OVMF.fd file, as bhyve will
map the bootrom file as readonly, and then a write will be
attempted.